pyproject.toml
[tool.poetry]
name = "iscc-service-generator"
version = "0.1.0"
description = "ISCC Generator Web Service APP"
authors = ["Titusz <tp@iscc.foundation>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/iscc/iscc-service-generator"
keywords=["iscc", "microservice", "rest-api", "identifier", "media", "hash", "simhash"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
iscc-sdk = {extras = ["turbo"], version = "^0.1.0"}
Django = "^4.0"
django-ninja = "^0.16"
django-admin-interface = "^0.18"
uvicorn = "^0.17"
django-model-utils = "^4.2"
blake3 = "^0.3."
humanize = "^3.14"
django-q = "^1.3"
django-constance = {extras = ["database"], version = "^2.8"}
django-json-widget = "^1.1"
django-object-actions = "^3.1"
dj-database-url = "^0.5"
pydantic = {extras = ["dotenv"], version = "1.9"}
psycopg2 = "^2.9"
requests = "^2.27"
pathvalidate = "^2.5"
translitcodec = "^0.7"


[tool.poetry.dev-dependencies]
black = "^22.1"
poethepoet = "^0.12"
mkdocs-material = "^8.1"
pytest = "^6.2"
pytest-django = "^4.5"
iscc-samples = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "iscc_service_generator.settings"

[tool.poe.tasks]
install = { cmd = "python -m dev.install", help = "Install app for development"}
